# backend/cmd/api/openapi.yaml
openapi: 3.1.0
info:
    title: University Room Booking API
    description: |
        REST API for the University Room Booking System.
        Allows users to search rooms, view availability timelines, create booking requests,
        and track their status. Administrators approve/reject pending bookings via a dashboard.

        ## Authentication
        - **Access Token**: Short-lived JWT (15 min) stored in an `httpOnly` cookie.
          `Authorization: Bearer <token>` remains supported for API clients.
        - **Refresh Token**: Long-lived token (30 days) stored in an `httpOnly` cookie with rotation on each use.

        ## Pagination
        All list endpoints use cursor-based pagination. The `nextCursor` value from a response
        should be passed as the `cursor` query parameter in the next request.

        ## Time Slots
        All booking times must be in 5-minute increments.
        Working day: 08:00–22:00.
    version: 1.0.0
    contact:
        name: Backend Architecture Team
servers:
    - url: http://localhost:3000/api
      description: Local development
    - url: https://booking.university.edu/api
      description: Production
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Compatibility access token auth for API clients
        AccessCookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
            description: httpOnly access token cookie
        RefreshCookieAuth:
            type: apiKey
            in: cookie
            name: refreshToken
            description: httpOnly refresh token cookie

    # ────────────────────────────────────────────
    # Reusable Schemas
    # ────────────────────────────────────────────
    schemas:
...
        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    type: object
                    required: [code, message]
                    properties:
                        code:
                            type: string
                            description: Machine-readable error code
                            example: BOOKING_CONFLICT
                        message:
                            type: string
                            description: Human-readable error description
                        details:
                            type: object
                            additionalProperties: true
...
paths:
...
    /media/rooms/{objectKey}:
        get:
            operationId: getRoomPhoto
            tags: [Rooms]
            summary: Get room photo
            description: Proxies a room photo stored in MinIO. Browser clients use this backend URL instead of direct MinIO access.
            parameters:
                - name: objectKey
                  in: path
                  required: true
                  schema:
                      type: string
                  description: Room photo object path, e.g. `{roomId}/{fileName}`.
            responses:
                "200":
                    description: Room photo bytes
                    content:
                        image/png:
                            schema:
                                type: string
                                format: binary
                        image/jpeg:
                            schema:
                                type: string
                                format: binary
                        image/webp:
                            schema:
                                type: string
                                format: binary
                "404":
                    description: Media not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
...